requirements.txt
torch==1.4.0
texar-pytorch==0.1.1